[Feat] Adds convert-moves app
This is still pretty hacky, but it generates GPX files that are good
enough for Strava.

So far, I've tested it on all the .sml files that I have, but I only
have twenty-eight of them.  My guess is they're deleted after a month.
Of the twenty-eight, I only uploaded the eight most recent ones to
Strava, so although there were no errors doing the conversion, there
could be bad being produced.