[Bug] no longer dies if a TrkPt can't be made.
My .sml file for today lacked speed on three samples.  I haven't seen
that before, but this commit makes it so that when we can't create a
TrkPt we just eprintln! a complaint rather than bubble up an error
that halts the program.

FWIW, the map that I get on Strava from today's GPX file is wrong in
the beginning, *but* when I view today's route on my watch, it's wrong
in the same way, so something weird was going on at the beginning of
my run.